Root finding results for bracketing method and secant method
For bracketing method, the output shows the values of a and b when they
converged to a value less than the level of tolerance, the actual root,
and the amount of iterations it took to converge to an answer. With the
secant method, it shows the last calculated approximation, the actual
root, and how many iterations it took to converge to an answer.